feat: support pure set and get
